requirements.txt
strawberry-graphql>=0.213.0
fastapi
uvicorn